# openapi.yaml
openapi: 3.0.3
info:
  title: FeedzzTrab API
  description: API para gerir atribuições de tarefas no sistema FeedzzTrab, projetado para gerentes, líderes de equipa e programadores.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor de desenvolvimento local
  - url: https://feedzztrab.example.com/api
    description: Servidor de produção (substitua pela sua URL alojada)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Utilizador:
      type: object
      properties:
        utilizador_id:
          type: integer
          example: 1
        nome_utilizador:
          type: string
          example: "programador1"
        email:
          type: string
          example: "programador1@exemplo.com"
        funcao:
          type: string
          enum: [admin, gerente, lider_equipa, programador]
          example: "programador"
        criado_em:
          type: string
          format: date-time
          example: "2025-03-22T10:00:00Z"
      required:
        - utilizador_id
        - nome_utilizador
        - email
        - funcao

    Competencia:
      type: object
      properties:
        competencia_id:
          type: integer
          example: 1
        nome_competencia:
          type: string
          example: "Desenvolvimento Web"
      required:
        - competencia_id
        - nome_competencia

    Tarefa:
      type: object
      properties:
        tarefa_id:
          type: integer
          example: 1
        descricao:
          type: string
          example: "Desenvolver um módulo de autenticação"
        criado_por:
          type: integer
          example: 2
        criado_em:
          type: string
          format: date-time
          example: "2025-03-22T10:00:00Z"
        estado:
          type: string
          enum: [nao_atribuida, atribuida, concluida]
          example: "nao_atribuida"
      required:
        - tarefa_id
        - descricao
        - criado_por
        - estado

    Atribuicao:
      type: object
      properties:
        atribuicao_id:
          type: integer
          example: 1
        tarefa_id:
          type: integer
          example: 1
        atribuido_a:
          type: integer
          example: 4
        atribuido_por:
          type: integer
          example: 3
        hora_inicio:
          type: string
          format: date-time
          nullable: true
          example: null
        hora_fim:
          type: string
          format: date-time
          nullable: true
          example: null
      required:
        - atribuicao_id
        - tarefa_id
        - atribuido_a
        - atribuido_por

    Erro:
      type: object
      properties:
        mensagem:
          type: string
          example: "Acesso não autorizado"
      required:
        - mensagem

security:
  - bearerAuth: []

paths:
  /utilizadores:
    post:
      summary: Criar um novo utilizador (apenas Admin)
      tags:
        - Utilizadores
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome_utilizador:
                  type: string
                  example: "programador2"
                palavra_passe:
                  type: string
                  example: "palavra_passe_segura123"
                email:
                  type: string
                  example: "programador2@exemplo.com"
                funcao:
                  type: string
                  enum: [admin, gerente, lider_equipa, programador]
                  example: "programador"
              required:
                - nome_utilizador
                - palavra_passe
                - email
                - funcao
      responses:
        '201':
          description: Utilizador criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilizador'
        '403':
          description: Proibido (não é admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /utilizadores/{utilizador_id}/competencias:
    post:
      summary: Adicionar uma competência ao perfil de um programador (apenas Programador)
      tags:
        - Competências
      security:
        - bearerAuth: []
      parameters:
        - name: utilizador_id
          in: path
          required: true
          schema:
            type: integer
            example: 4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                competencia_id:
                  type: integer
                  example: 2
              required:
                - competencia_id
      responses:
        '201':
          description: Competência adicionada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competencia'
        '403':
          description: Proibido (não é o programador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /utilizadores/{utilizador_id}/palavra_passe:
    patch:
      summary: Alterar a palavra-passe do utilizador (qualquer utilizador autenticado)
      tags:
        - Utilizadores
      security:
        - bearerAuth: []
      parameters:
        - name: utilizador_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                palavra_passe_atual:
                  type: string
                  example: "palavra_passe123"
                nova_palavra_passe:
                  type: string
                  example: "nova_palavra_passe456"
              required:
                - palavra_passe_atual
                - nova_palavra_passe
      responses:
        '200':
          description: Palavra-passe alterada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Palavra-passe alterada com sucesso"
        '400':
          description: Palavra-passe atual incorreta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Proibido (não é o próprio utilizador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /tarefas:
    post:
      summary: Criar uma nova tarefa (apenas Gerente)
      tags:
        - Tarefas
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                descricao:
                  type: string
                  example: "Corrigir erro na página inicial"
              required:
                - descricao
      responses:
        '201':
          description: Tarefa criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
        '403':
          description: Proibido (não é gerente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

    get:
      summary: Listar tarefas (Gerente vê as suas, Líder de Equipa vê todas)
      tags:
        - Tarefas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tarefas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tarefa'
        '403':
          description: Proibido (não é gerente nem líder de equipa)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /tarefas/{tarefa_id}:
    put:
      summary: Atualizar uma tarefa não atribuída (apenas Gerente)
      tags:
        - Tarefas
      security:
        - bearerAuth: []
      parameters:
        - name: tarefa_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                descricao:
                  type: string
                  example: "Atualizar módulo de autenticação"
              required:
                - descricao
      responses:
        '200':
          description: Tarefa atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
        '403':
          description: Proibido (não é gerente ou tarefa já atribuída)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

    delete:
      summary: Eliminar uma tarefa não atribuída (apenas Gerente)
      tags:
        - Tarefas
      security:
        - bearerAuth: []
      parameters:
        - name: tarefa_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Tarefa eliminada com sucesso
        '403':
          description: Proibido (não é gerente ou tarefa já atribuída)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /atribuicoes:
    post:
      summary: Atribuir uma tarefa a um programador (apenas Líder de Equipa)
      tags:
        - Atribuições
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tarefa_id:
                  type: integer
                  example: 1
                atribuido_a:
                  type: integer
                  example: 4
              required:
                - tarefa_id
                - atribuido_a
      responses:
        '201':
          description: Tarefa atribuída com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Atribuicao'
        '403':
          description: Proibido (não é líder de equipa)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '409':
          description: Conflito (programador tem uma tarefa ativa)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /atribuicoes/{atribuicao_id}:
    patch:
      summary: Atualizar horário de início/fim de uma atribuição (apenas Programador)
      tags:
        - Atribuições
      security:
        - bearerAuth: []
      parameters:
        - name: atribuicao_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hora_inicio:
                  type: string
                  format: date-time
                  nullable: true
                  example: "2025-04-01T09:00:00Z"
                hora_fim:
                  type: string
                  format: date-time
                  nullable: true
                  example: "2025-04-01T17:00:00Z"
      responses:
        '200':
          description: Atribuição atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Atribuicao'
        '403':
          description: Proibido (não é programador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'